requirements.txt
# tensorflow
# keras
# numpy
# matplotlib
# pygame
# opencv-python
torch
pillow
torchvision
torchaudio
